+++ example/macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -7,7 +7,6 @@ export "FLUTTER_TARGET=/Users/rafaelbarbosa/Documents/MyProjects/bonfire/example
 export "FLUTTER_BUILD_DIR=build"
 export "FLUTTER_BUILD_NAME=1.0.0"
 export "FLUTTER_BUILD_NUMBER=1"
-export "EXCLUDED_ARCHS=arm64"
 export "DART_DEFINES=Zmx1dHRlci5pbnNwZWN0b3Iuc3RydWN0dXJlZEVycm9ycz10cnVl,RkxVVFRFUl9XRUJfQVVUT19ERVRFQ1Q9dHJ1ZQ=="
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"